[106] - some test fixes; should fix GetAllLotsByLocationTests (comments left)
